move lobby navbar to component
